pyproject.toml
[project]
name = "qr-debug-camera"
version = "0.1.0"
description = "Screen QR reader that injects a debug camera stream into Chrome."
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "mss>=9.0.1",
  "numpy>=1.26",
  "opencv-python>=4.10",
  "PySide6>=6.7",
  "tzdata>=2024.1",
  "websocket-client>=1.8",
]

[project.scripts]
qr-debug-camera = "qr_debug_camera.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
  "mypy>=1.10",
  "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
warn_unused_ignores = true
check_untyped_defs = true
